Rename to paddleaudio.
