+++ paddleaudio/examples/panns/audio_tag.py
@@ -17,6 +17,7 @@ from typing import List
 
 import numpy as np
 import paddle
+
 from paddleaudio.backends import load as load_audio
 from paddleaudio.features import melspectrogram
 from paddleaudio.models.panns import cnn14
