+++ paddleaudio/examples/panns/parse_result.py
@@ -17,6 +17,7 @@ import os
 from typing import Dict
 
 import numpy as np
+
 from paddleaudio.utils import logger
 
 # yapf: disable
