+++ paddleaudio/examples/sound_classification/deploy/python/predict.py
@@ -16,10 +16,11 @@ import os
 
 import numpy as np
 from paddle import inference
+from scipy.special import softmax
+
 from paddleaudio.backends import load as load_audio
 from paddleaudio.datasets import ESC50
 from paddleaudio.features import melspectrogram
-from scipy.special import softmax
 
 # yapf: disable
 parser = argparse.ArgumentParser()
